requirements.txt
groq>=0.13.0
python-dotenv>=1.0.0
